+++ sdk/eventhub/event-hubs/src/eventDataAdapter.ts
@@ -10,11 +10,11 @@ import { EventData } from "./eventData";
  *
  * @hidden
  */
-export interface MessageWithMetadata {
+export interface MessageContent {
   /**
    * The message's binary data
    */
-  body: Uint8Array;
+  data: Uint8Array;
   /**
    * The message's content type
    */
@@ -33,11 +33,11 @@ export interface MessageAdapter<MessageT> {
   /**
    * defines how to create a message from a payload and a content type
    */
-  produceMessage: (messageWithMetadata: MessageWithMetadata) => MessageT;
+  produceMessage: (MessageContent: MessageContent) => MessageT;
   /**
    * defines how to access the payload and the content type of a message
    */
-  consumeMessage: (message: MessageT) => MessageWithMetadata;
+  consumeMessage: (message: MessageT) => MessageContent;
 }
 
 // This type should always be equivalent to Omit<Omit<EventData, "body">, "contentType">
@@ -79,17 +79,17 @@ export function createEventDataAdapter(
   params: EventDataAdapterParameters = {}
 ): MessageAdapter<EventData> {
   return {
-    produceMessage: ({ body, contentType }: MessageWithMetadata) => {
+    produceMessage: ({ data: body, contentType }: MessageContent) => {
       return {
         ...params,
         body,
         contentType,
       };
     },
-    consumeMessage: (message: EventData): MessageWithMetadata => {
+    consumeMessage: (message: EventData): MessageContent => {
       const { body, contentType } = message;
       if (body === undefined) {
-        throw new Error("Expected the body field to be defined");
+        throw new Error("Expected the data field to be defined");
       }
       if (contentType === undefined) {
         throw new Error("Expected the contentType field to be defined");
@@ -99,7 +99,7 @@ export function createEventDataAdapter(
          * If the raw response was parsed as JSON, we need to convert it to a Uint8Array,
          * otherwise, leave the payload as is.
          */
-        body: typeof body === "object" ? Uint8Array.from(Object.values(body)) : body,
+        data: typeof body === "object" ? Uint8Array.from(Object.values(body)) : body,
         contentType,
       };
     },